+++ ark-admin/src/main/java/au/org/theark/admin/service/IAdminService.java
@@ -92,15 +92,15 @@ public interface IAdminService<T> {
 
 	public long getArkModuleCount(ArkModule arkModuleCriteria);
 
-	public List<ArkModule> searchPageableArkModules(ArkModule arkModuleCriteria, int first, int count);
+	public List<ArkModule> searchPageableArkModules(ArkModule arkModuleCriteria, long first, long count);
 
 	public long getArkFunctionCount(ArkFunction arkFunctionCriteria);
 
-	public List<ArkFunction> searchPageableArkFunctions(ArkFunction arkFunctionCriteria, int first, int count);
+	public List<ArkFunction> searchPageableArkFunctions(ArkFunction arkFunctionCriteria, long first, long count);
 
 	public long getArkRoleModuleFunctionVOCount(ArkRoleModuleFunctionVO arkRoleModuleFunctionVO);
 	
-	public List<ArkRoleModuleFunctionVO> searchPageableArkRoleModuleFunctionVO(ArkRoleModuleFunctionVO arkRoleModuleFunctionVo, int first, int count);
+	public List<ArkRoleModuleFunctionVO> searchPageableArkRoleModuleFunctionVO(ArkRoleModuleFunctionVO arkRoleModuleFunctionVo, long first, long count);
 	
 	public ArkRole getArkRoleByName(String name);
 
@@ -110,7 +110,7 @@ public interface IAdminService<T> {
 
 	public long getArkModuleFunctionCount(ArkModuleFunction arkModuleFunction);
 	
-	public List<ArkModuleFunction> searchPageableArkModuleFunctions(ArkModuleFunction arkModuleFunctionCriteria, int first, int count);
+	public List<ArkModuleFunction> searchPageableArkModuleFunctions(ArkModuleFunction arkModuleFunctionCriteria, long first, long count);
 
 	public ArkModuleFunction getArkModuleFunction(Long id);
 
@@ -122,7 +122,7 @@ public interface IAdminService<T> {
 	
 	public long getArkRoleCount(ArkRole arkRoleCriteria);
 
-	public List<ArkRole> searchPageableArkRoles(ArkRole arkRoleCriteria, int first, int count);
+	public List<ArkRole> searchPageableArkRoles(ArkRole arkRoleCriteria, long first, long count);
 	
 	public void createOrUpdateArkRole(AdminVO modelObject);
 
@@ -132,7 +132,7 @@ public interface IAdminService<T> {
 	
 	public long getArkModuleRoleCount(ArkModuleRole arkModuleRole);
 	
-	public List<ArkModuleRole> searchPageableArkModuleRoles(ArkModuleRole arkModulRoleCriteria, int first, int count);
+	public List<ArkModuleRole> searchPageableArkModuleRoles(ArkModuleRole arkModulRoleCriteria, long first, long count);
 
 	public void createArkModuleRole(AdminVO modelObject);
