+++ ark-disease/src/main/java/au/org/theark/disease/dao/IDiseaseDao.java
@@ -1,7 +1,6 @@
 package au.org.theark.disease.dao;
 
 import java.util.List;
-import java.util.Set;
 
 import au.org.theark.core.model.disease.entity.Affection;
 import au.org.theark.core.model.disease.entity.AffectionCustomFieldData;
@@ -11,7 +10,6 @@ import au.org.theark.core.model.disease.entity.Gene;
 import au.org.theark.core.model.disease.entity.Position;
 import au.org.theark.core.model.study.entity.LinkSubjectStudy;
 import au.org.theark.core.model.study.entity.Study;
-import au.org.theark.core.vo.CustomFieldVO;
 import au.org.theark.disease.vo.AffectionListVO;
 import au.org.theark.disease.vo.AffectionVO;
 import au.org.theark.disease.vo.DiseaseVO;
@@ -21,11 +19,11 @@ public interface IDiseaseDao {
 	
 	public int getDiseaseCount(DiseaseVO diseaseVO);
 	
-	public List<DiseaseVO> searchPageableDiseases(DiseaseVO object, int first, int count);
+	public List<DiseaseVO> searchPageableDiseases(DiseaseVO object, long first, long count);
 	
 	public int getGeneCount(GeneVO geneVO);
 
-	public List<GeneVO> searchPageableGenes(GeneVO object, int first, int count);
+	public List<GeneVO> searchPageableGenes(GeneVO object, long first, long count);
 	
 	public void save(Object object);
 	
@@ -39,7 +37,7 @@ public interface IDiseaseDao {
 
 	public int getAffectionCount(AffectionVO affectionVO);
 
-	public List<AffectionVO> searchPageableAffections(AffectionVO affectionVO, int first, int count);
+	public List<AffectionVO> searchPageableAffections(AffectionVO affectionVO, long first, long count);
 	
 	public List<AffectionStatus> getAffectionStatus();
