+++ ark-disease/src/main/java/au/org/theark/disease/web/component/disease/DiseaseContainerPanel.java
@@ -1,9 +1,13 @@
 package au.org.theark.disease.web.component.disease;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-
+import au.org.theark.core.Constants;
+import au.org.theark.core.model.study.entity.Study;
+import au.org.theark.core.service.IArkCommonService;
+import au.org.theark.core.web.component.AbstractContainerPanel;
+import au.org.theark.core.web.component.ArkDataProvider;
+import au.org.theark.disease.service.IArkDiseaseService;
+import au.org.theark.disease.vo.DiseaseVO;
+import au.org.theark.disease.web.component.disease.form.ContainerForm;
 import org.apache.shiro.SecurityUtils;
 import org.apache.wicket.ajax.AjaxRequestTarget;
 import org.apache.wicket.ajax.markup.html.navigation.paging.AjaxPagingNavigator;
@@ -15,14 +19,9 @@ import org.apache.wicket.spring.injection.annot.SpringBean;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import au.org.theark.core.Constants;
-import au.org.theark.core.model.study.entity.Study;
-import au.org.theark.core.service.IArkCommonService;
-import au.org.theark.core.web.component.AbstractContainerPanel;
-import au.org.theark.core.web.component.ArkDataProvider;
-import au.org.theark.disease.service.IArkDiseaseService;
-import au.org.theark.disease.vo.DiseaseVO;
-import au.org.theark.disease.web.component.disease.form.ContainerForm;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 
 public class DiseaseContainerPanel extends AbstractContainerPanel<DiseaseVO> {
 
@@ -95,11 +94,11 @@ public class DiseaseContainerPanel extends AbstractContainerPanel<DiseaseVO> {
 		diseaseProvider = new ArkDataProvider<DiseaseVO, IArkDiseaseService>(iArkDiseaseService) {
 			private static final long serialVersionUID = 1L;
 			
-			public int size() {
+			public long size() {
 				return service.getDiseaseCount(containerForm.getModelObject());
 			}
 
-			public Iterator<? extends DiseaseVO> iterator(int first, int count) {
+			public Iterator<? extends DiseaseVO> iterator(long first, long count) {
 				List<DiseaseVO> listDiseases = new ArrayList<DiseaseVO>();
 				listDiseases = service.searchPageableDiseases(model.getObject(), first, count);
 				return listDiseases.iterator();
